Added very basic RSpec rig to the MadChatter project.
Among the many, many testing libraries for Ruby rspec is one of the
more popular, with an active community of folks pumping out code so
you don't have to (until you do). For instance, Guard is a related
tool that will use filesystem notifications to re-run certain
commands; the Guardfile in this project will now run specs when
source-code and spec-code is changed. Just run 'guard' in the
top-level of the project.

The actual spec/message_spec.rb is not too bright, but it's a
start. No attempt is made at integration testing, though when
EventMachine has to be wrangled consider:

  * perfect black-box browser based testing (watir, etc.)
  * evented-spec with em-synchronity
  * petitioning the lord with prayer

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>